refactor(system-config): move LocaleTypeHandler to THL
